Move shared logging data to the desktop-shared library
