+++ libs/desktop-shared/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/interfaces';
+export * from './lib';
